Move LiveKit module into its own directory
